refactor: spring boot data rest 의존성 삭제
