+++ src/main/java/com/twohoseon/app/controller/AuthRestController.java
@@ -1,8 +1,8 @@
 package com.twohoseon.app.controller;
 
-import com.twohoseon.app.dto.GeneralResponseDTO;
-import com.twohoseon.app.dto.TokenDTO;
-import com.twohoseon.app.dto.TokenRefreshDTO;
+import com.twohoseon.app.dto.request.TokenRefreshDTO;
+import com.twohoseon.app.dto.response.GeneralResponseDTO;
+import com.twohoseon.app.dto.response.TokenDTO;
 import com.twohoseon.app.enums.StatusEnum;
 import com.twohoseon.app.service.refreshToken.RefreshTokenService;
 import lombok.RequiredArgsConstructor;
