+++ src/main/java/com/twohoseon/app/security/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.twohoseon.app.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.twohoseon.app.dto.GeneralResponseDTO;
+import com.twohoseon.app.dto.response.GeneralResponseDTO;
 import com.twohoseon.app.entity.Member;
 import com.twohoseon.app.enums.StatusEnum;
 import com.twohoseon.app.enums.UserRole;
