+++ src/main/java/com/twohoseon/app/service/refreshToken/RefreshTokenService.java
@@ -1,6 +1,6 @@
 package com.twohoseon.app.service.refreshToken;
 
-import com.twohoseon.app.dto.TokenDTO;
+import com.twohoseon.app.dto.response.TokenDTO;
 
 /**
  * @author : hyunwoopark
